requirements.txt
selenium==3.11.0
python-dotenv==0.8.2
chromedriver_installer===0.0.6
Jinja2===2.10